Fix synthesis race condition (#17)
Fixes two race conditions: one in the test and one in the code.

The main race condition is reached when a composition has been created
but not yet synthesized. It will be eligible for rollout, even though
that doesn't make sense. We will always reach it on the next pass of the
loop anyway.

The test race condition is simpler. It occurs because we use two
separate test compositions in the test, and Kubernetes doesn't provide
synchronization across resources. The same race isn't an issue in
production because we don't care so much about ordering in within the
eventually consistent system.

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>