Reconciliation controller (#20)
Adds an initial implementation of the reconciliation controller with
integration tests that run against a matrix of all Kubernetes versions
supported by envtest (which is a lot 🙃 ).

Deletes aren't implemented yet, since we still need to implement
tombstones. And there are still some rough edges in the test
infrastructure. Cutting a PR now to avoid it getting any larger.

---------

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>