Add synthesizer execution controller (#22)
The exec controller is responsible for actually invoking the synthesizer
processes within their containers.

Wiring this up involves a fair number of changes to the existing
synthesis controllers because the semantics have changed somewhat:
previously they assumed the container processes were one-shots that exit
0 once the resources have been synthesized. Now the container process
just sleeps and we exec into the container to spawn the actual
synthesizer. This approach is nice because it doesn't require a wrapper
process like the prototype did.

This PR has minimal tests- we'll need to add more for the exec
controller later. Just trying to keep the PR a reasonable length.

---------

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>