Fix informer race condition (#73)
This fixes an exceptionally rare test flake in which the status of a
resource isn't ever populated. It's caused by the status being flushed
out of the write buffer before the corresponding resource slice is
cached in the informer. This is possible because the reconstitution
cache doesn't read from the informer cache. So it's possible to get a
composition change event, fill the cache, and reconcile a resource
before the resource slice hits the informer.

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>