More symphony-related improvements (#103)
- Terminating pods shouldn't block composition deletion
- Symphony status fields should be nil if _any_ of the compositions have
not reached that state. Currently it takes the max value.
- Don't re-delete resources in a couple of obscure cases
- Retry stale cache errors

That last one is tricky. When reading compositions from the resource
slice status controller we error on 404s within 1min of resource slice
creation. This is an excessively long time for informers to be sync'd,
typically another event is received within a couple of ms. But there is
one case in which this code can be reached legitimately: compositions
that were deleted during synthesis. A resource slice may have already
been written before deletion, so it will (rightly) not hold the
composition finalizer. It's fine to defer GC'ing that slice for 1min but
currently we never retry, so it will be blocked until resync.

---------

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>