Don't create compositions for symphonies in terminating namespaces (#132)
When a namespace is deleted, we might still process a symphony before
the informer catches up and the deletion timestamp on it is updated. So
we might try to re-create compositions as they're being deleted and
retry on the errors. This change makes it so we drop the queue item
instead of retrying when encountering a terminating namespace.

---------

Co-authored-by: Mariano Uvalle <juvallegarza@microsoft.com>
Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>