Reduce conflicts (#138)
There are two controllers that often throw conflict errors since they
update resources concurrent to other controllers. In one case there is
no risk of meaningful conflict so we can just patch. In the other case
(status controller) there is technically a condition in which the
ordering matters so we'll have to retry conflicts. This could be
refactored out in the future.

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>