Better orphan'd resource cleanup (#192)
Currently Eno will recreate missing namespaces in order to finish
deleting any symphonies that still reference them. This behavior was
symphony specific, which kind of makes sense given its role in the Eno
ecosystem.

The problem is resources that _could in the future be_ related to a
symphony, but are not yet. Specifically: resource slices. It's possible
that synthesis was in-progress when the namespace was force-deleted, and
that resource slices have been written but not yet associated with the
composition.

I propose that we just enable the cleanup logic for all resources. There
isn't any real reason for it to be exclusive to symphonies - just try to
recreate the namespace any time it's still missing after 1 second.

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>