Merge watchdog controller into scheduling controller (#283)
Both controllers respond to essentially every Eno CR state transition,
so merging the controllers will help reduce overhead.

This does effectively remove a few metrics that we haven't found useful.
The watchdog controller is profoundly useful for the one metric that we
are keeping, but the others are not well-scoped so the logs are better
anyway.

---------

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>