More intuitive latency reporting for write buffer (#290)
Currently the write buffer reports the latency as the period between the
latest item being added to the buffer and it being flushed to the
server.

This doesn't map to what I (most people?) would expect, which is the
time between the _first_ item being added to the buffer and it being
flushed.

So the fix is to just make the insertion time follow the queue item
lifecycle: set it when initially enqueuing, remove it when forgetting
the rate limit.

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>