Handle case where resource version doesn't change after update (#292)
The builtin APIs are full of edge cases where SMD and even the client-go
scheme are not capable of comparing two versions of an API object. A few
examples are documented in integration tests: resource quantities,
fieldRefs, etc.

Eno currently uses a combination of SMD's equality logic and the scheme.
But it turns out that the client-go scheme doesn't support defaulting -
that requires importing a scheme for kubernetes/kubernetes, which is...
not ideal.

So this PR proposes a different approach: keep updating until the
resource version doesn't change. This essentially outsources the
equality logic to apiserver: it compares the current/updated objects to
determine if the write (and thus new resource version) is necessary.

Ideally synthesizers would set the relevant fields to avoid hitting this
code. But IMO it's a better protection against tightloops than fully
trusting Eno's comparison logic.

---------

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>