Add fuzz-style test framework (#297)
Eno already has some property-based testing (style) tests in the
scheduling package, but there's a lot of boilerplate involved and some
indirection would be helpful if we're going to expand the number of
tests significantly.

This framework essentially just re-implements the existing boilerplate
but slightly nicer. I'll update the test case to use it in another PR.

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>