Extract pod garbage collection into a separate controller (#299)
Currently the synthesis lifecycle controller creates and deletes pods.
But since it reconciles compositions, it needs to reconcile the entire
set of pods associated with the composition for every reconciliation.
This works, but it's confusing to read/test, and it makes requeing to
check timeouts tricky.

This PR refactors the deletion logic into a separate controller, extends
test coverage around some important edge cases, and simplifies the
existing lifecycle controller somewhat. Mostly it unblocks better
testing of the lifecycle controller in the future.

---------

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>