Allow election retry period to be set (#305)
The current election defaults are a bit lazy - waiting 90 seconds after
the leader dies is not great. This PR adopts the values used by the
various Flux controllers which have been proven to work well in most
cases.

The retry period is exposed as a flag now, which is useful for reducing
request volume from eno-reconciler processes in cases where many may be
running concurrently.

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>