Handle missing slices for previous synthesis (#314)
Currently, reconciliation will eventually converge after deleting the
resource slices out from under a composition - but only if the
composition does not have a previous synthesis. (i.e. it was only
synthesized once)

The fix is simple: bail out of the cache controller when a slice is
missing. This is safe even if compositions/slices hit the informers out
of order since the controller watches both types, so we'll always get
another workqueue item eventually.

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>