Improve error logging fields (#319)
Currently errors are wrapped and returned by the controller, and logged
by controller-runtime internals. This is nice since it avoids the common
issue of logging the same error multiple times, but it doesn't provide a
good way to include the log fields that have accumulated on the logger
through the Reconcile call.

I think we should break down and just log the errors from the Reconcile
function to preserve the logger fields.

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>