Handle slow synthesis pod deletion (#329)
Currently replacement synthesis pods are only created after the existing
pods are fully deleted i.e. gone from the informer entirely. The
intention of the design was to create a new pod as soon as the old pod
is marked for deletion, so we need to also enqueue when the deletion
timestamp transitions from nil to non-nil.

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>