Server-side apply (#337)
I think it's time to accept defeat on client-side apply. 😢 

The structured-merge-diff library is great, except it doesn't (well,
can't) apply default values. So the current approach works fine until an
array object has a `x-kubernetes-list-map-keys` that doesn't match its
`x-kubernetes-patch-merge-key`.

For example: the service port map key is protocol+port, but its merge
key is only the port. Objects that set both properties can be reconciled
by Eno. But setting only the port will fail, even though it would
succeed with client-side three-way merge. This is because `protocol`
defaults to TCP, but that default is applied in apiserver - not SMD.

We can keep hacking around this by modifying the spec, but the reasons
to avoid SSA have been exhausted at this point (I'll avoid enumerating
them here).

---------

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>