fix(helmshim): skip nil helm object (#366)
Helm renderer can generate empty yaml file like this:

```yaml
---
# source: a.yaml
# comment
---
# source: b.yaml
apiVersion: v1
kind: ConfigMap
# ... object data
```

In this case, the section of a.yaml will be decoded as an empty
`unstructured.Unstructred` object. However, object without GVK is not
supported in eno and it will be rejected in the output writer:
https://github.com/Azure/eno/blob/07e348b023dc93833fb509ff6f8aaa2b014e173a/pkg/function/outputs.go#L56-L63

But in current helmshim implementation, the return value of output
writer `Write` is ignored, hence it suppressed the underlying issue.

This pull request addressed this issue by:

1. handle error from output writer `Write` method
2. skip helm objects that rendered as empty `unstructured.Unstructured`
object by checking on the JSON representation.