Don't reconcile resource once a matching patch has been applied (#389)
There's an obscure case where patches don't behave as expected: if a
patch is defined for a resource that is also managed by Eno, and the
patch is in a later readiness group, both logical resources will be
reconciled normally, which can result in oscillation between states.

The fix is to skip the earlier resource once the patch has been applied.

I'm not sure this has real consequence for any reasonable use case, but
it's causing test flakes and worth fixing anyway.

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>