Sync composition status when synthesizer changes (#391)
Currently the controller doesn't always transition in/out of the
`MissingSynthesizer` phase correctly, since there is no guarantee that
the status is updated after the synthesizer is created or deleted.

This is strictly a UX bug since there it doesn't impact the behavioral
outside of the reported composition status. But it does sometimes cause
test flakes for the missing input tests.

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>