Fix --disable-ssa (#396)
Resources that are modified by other clients will not currently work
correctly when `--disable-ssa` is set because it computes the merge
patch based on the current state - not the previous. So any properties
set outside of Eno will be removed during reconciliation.

This fixes the bug by doing `s/current/previous/g` and adds smoke test
coverage.

---------

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>