Don't compare timestamps when checking fix missing managedFields (#399)
Currently the `managedFields` metadata backfilling logic only works when
the time between the last update to the managedFields and the current
time (during reconciliation) is <1s. It was overlooked because the
reconciliation controller is generally pretty fast - it's only an issue
when apiserver is loaded down and slow to respond.

This PR fixes the bug by only comparing the fields themselves, which
naturally excludes the last update timestamp from the comparison.

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>