Introduce resource.Snapshot (#416)
Small refactor to enable an upcoming change.

The idea is to more explicitly capture the expected state of a resource
at a particular point in (resourceVersion) time given its overrides.
Currently we just apply the overrides when they're needed, which is
surprisingly difficult to reason about and in some cases generates
unnecessary garbage.

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>