Simplify synthesis timeouts (#465)
Currently each synthesizer can define its own `execTimeout` and
`podTimeout`. Exec timeouts apply when the eno-executor process shells
out to the synthesizer, pod timeouts are the max lifespan of each
synthesizer pod object.

It doesn't make sense for pod timeouts to be defined by the synthesizer
since they don't do IO. The timeout value is a product of the platform
(pod creation time) which is generally owned by a different persona than
that of most synthesizers.

Similarly, execTimeout is useless since the processes don't do any IO.
In real world use-cases we've never actually reached this timeout.

---

So let's remove the exec timeout and move the pod timeout to a CLI flag
instead of synthesizer field. Marking the fields as deprecated for now,
will remove in a future release.

---------

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>