Tolerate nil current object in CEL runtime (#474)
Currently the `internal/cel` package assumes that the current object is
non-nil. There are checks in the caller to guarantee this, but the
behavior doesn't actually make sense - users may want conditions that
match on the composition metadata regardless of if the resource already
exists.

So this change just updates the cel package in anticipation of the
caller-side change.

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>