Support per-resource deletion strategy (#487)
It's hard to know which annotations can be set on compositions vs
individual resources.

Currently the worst offender is `deletion-strategy` which is
surprisingly on the composition. This PR moves it to the resource and
updates all "snapshotted" (i.e. overridable) annotations to use
cascading reads: the resource's value is used if set, otherwise the
value is inherited from the composition.

---------

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>