Enable transferring resources between compositions (#494)
Overrides already have the primitives needed to gracefully transfer the
management of resources between synthesizers/compositions using either
composition or resource annotations. This PR adds test coverage and
makes a couple of small fixes that enable the workflow.

- Honor the `eno.azure.io/reconcile-interval` annotation when
`eno.azure.io/disable-reconciliation == 'true'`: this is necessary for
cases where the composition receiving ownership of an existing resource
is not resynthesized during the migration
- Allow overrides to be applied during initial resource creation: it
made sense to block overrides during creation initially since the only
value available to override expressions was the current state of the
resource, which would naturally be nil. But now that we support
composition metadata it doesn't make sense and complicates the resource
transfer workflow.
- Fix the delete composition resource orphaning test: it was broken
because the override condition now blocks initial resource creation
(this is a very uncommon and recently added scenario and shouldn't cause
actual backcompat issues)

---------

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>