Expose reconciliation errors (#506)
Sometimes it doesn't make sense to sift through the eno-reconciler logs
to uncover why a composition never exits the `Reconciling` state e.g.
when a synthesizer generates an invalid resource, or a CR when the CRD
is missing.

This change simplifies the message and passes it through the resource
slice into the composition's simplified status.

---------

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>