Add readiness logic for resource deletion path (#514)
The resource readiness signal isn't currently used when deleting
resources since readiness expressions shouldn't apply to deleting
resources.

But in order to implement ordered deletion we need some knowledge of
when the resource is fully deleted e.g. returns 404 vs a 200 response
with `deletionTimestamp !=nil`. We can easily reuse the concept of
readiness to track the deletion state for this purpose.

This PR also refactors the tests to better for the readiness path - some
only wait on reconciliation. All of the logic in `checkReadiness` that
is important to current behaviors has test coverage and the new branches
will be covered more directly in the upcoming deletion order integration
tests.

---------

Co-authored-by: Jordan Olshevski <joolshev@microsoft.com>