add optional refs for synthesizer (#541)
Add optional field for refs in synthesizer. Default is false where
synthesis will be blocked if input is missing. If this field is set to
true, then synthesis will be non-blocking if input is missing

considers cases for removing revisions when the optional ref/input is
deleted.
default behavior where ref is required should still produce missing
inputs error

## Testing Done
* [x] unit tests
* [x] integraiton tests
* [ ] manual testing